+++ dialect/mysql/dialect/mods.go
@@ -328,12 +328,20 @@ func (l LockChain[Q]) SkipLocked() LockChain[Q] {
 	})
 }
 
-type WindowMod[Q interface{ AppendWindow(clause.NamedWindow) }] struct {
-	Name string
+type WindowMod[Q interface{ SetWindow(clause.Window) }] struct {
 	*WindowChain[*WindowMod[Q]]
 }
 
 func (w WindowMod[Q]) Apply(q Q) {
+	q.SetWindow(w.def)
+}
+
+type WindowsMod[Q interface{ AppendWindow(clause.NamedWindow) }] struct {
+	Name string
+	*WindowChain[*WindowsMod[Q]]
+}
+
+func (w WindowsMod[Q]) Apply(q Q) {
 	q.AppendWindow(clause.NamedWindow{
 		Name:       w.Name,
 		Definition: w.def,
@@ -341,7 +349,7 @@ func (w WindowMod[Q]) Apply(q Q) {
 }
 
 type WindowChain[T any] struct {
-	def  clause.WindowDef
+	def  clause.Window
 	Wrap T
 }